% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_light_threshold.R
\name{area_light_temp_threshold}
\alias{area_light_temp_threshold}
\title{Calculate the average benthic area within temperature and light threshold}
\usage{
area_light_temp_threshold(wtr, kd, light_incident, irr_thresh = c(0, 2000),
  wtr_thresh = c(0, 25), hypso, area_type = "surface")
}
\arguments{
\item{wtr}{Data frame of water temperatures. See \code{\link[glmtools]{get_temp}}}

\item{kd}{Light attenuation value of the lake (in m^-1 units)}

\item{light_incident}{Vector of incident light (arbitrary units but units must match threshold units).}

\item{irr_thresh}{Vector of length 2 with min and max value (in that order) for light thresholds. Note: 
>= and <= is used for comparison, so thresholds must be adjusted accordingly.}

\item{wtr_thresh}{Vector of length 2 with min and max value (in that order) for temperature thresholds. Note: 
>= and <= is used for comparison, so thresholds must be adjusted accordingly.}

\item{hypso}{Hypsography data.frame, with columns \code{depth} and \code{area}. From \code{\link{}}}

\item{area_type}{How benthic area is to be calculated. Defaults to \code{benthic}, which tries to estimate the
slope corrected benthic area for each layer. \code{surface} just uses a surface area approximation, 
ignoring slope.}
}
\author{
Luke Winslow
}

