% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_light_threshold.R
\name{area_light_threshold}
\alias{area_light_threshold}
\title{Calculate the average benthic area within light thresholds}
\usage{
area_light_threshold(kd, light_incident, irr_thresh = c(0, 2000), hypso,
  area_type = "benthic")
}
\arguments{
\item{kd}{Light attenuation value of the lake (in m^-1 units)}

\item{light_incident}{Vector of incident light (arbitrary units but units must match threshold units).}

\item{irr_thresh}{Vector of length 2 with min and max value (in that order) for light thresholds. Note: 
>= and <= is used for comparison, so thresholds must be adjusted accordingly.}

\item{hypso}{Hypsography data.frame, with columns \code{depth} and \code{area}. From \code{\link{}}}

\item{area_type}{How benthic area is to be calculated. Defaults to \code{benthic}, which tries to estimate the
slope corrected benthic area for each layer. \code{surface} just uses a surface area approximation, 
ignoring slope.}
}
\description{
Returns the average amount of benthic area that is within the optical intensity thresholds based on 
light attenuation, hypsograph, and the desired threshold values.
}
\author{
Luke Winslow
}
