% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/volume_light_threshold.R
\name{area_temp_threshold}
\alias{area_temp_threshold}
\title{Calculate the average benthic area within temperature thresholds}
\usage{
area_temp_threshold(wtr, wtr_thresh = c(0, 25), hypso,
  area_type = "surface")
}
\arguments{
\item{wtr}{Data frame of water temperatures. See \code{\link[glmtools]{get_temp}}}

\item{wtr_thresh}{Vector of length 2 with min and max value (in that order) for temperature thresholds. Note:
>= and <= is used for comparison, so thresholds must be adjusted accordingly.}

\item{hypso}{Hypsography data.frame, with columns \code{depth} and \code{area}. From \code{\link{}}}

\item{area_type}{How benthic area is to be calculated. Defaults to \code{benthic}, which tries to estimate the
slope corrected benthic area for each layer. \code{surface} just uses a surface area approximation,
ignoring slope.}
}
\description{
Calculate the average benthic area within temperature thresholds
}
\author{
Luke Winslow
}

