% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mod_obs_metric.R
\name{calc_mod_obs_metric}
\alias{calc_mod_obs_metric}
\title{Calculate limno metrics for lots of lakes}
\usage{
calc_mod_obs_metric(mod_obs_df, metric)
}
\description{
Calculte aggregate metrics from a data frame with modeled and observed data. 
Metric options include all available rLakeAnalyzer metrics
}
\seealso{
\code{\link[glmtools]{compare_to_field}}
}

