% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stratified_periods.R
\name{calc_stratified_periods}
\alias{calc_stratified_periods}
\title{Calculate all stratified periods based on temp threshold}
\usage{
calc_stratified_periods(surfT, botT, temp_thresh = 1, force_positive = TRUE)
}
\arguments{
\item{surfT}{data.frame with two columns, datetime and surface temperature}

\item{botT}{data.frame with two columns, datetime and bottom temperature (must be same nrow as surfT)}

\item{temp_thresh}{Threshold to define stratified in deg C (default = 1)}

\item{force_positive}{Is only warm stratified periods to be considered (drop winter when 0 is above 4 deg water, default TRUE)}
}
\description{
A function to calculate
}

