% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_toha.R
\name{calc_toha}
\alias{calc_toha}
\title{calculate the Thermal Optical Habitat Area (TOHA) for a given lake}
\usage{
calc_toha(lakeid, secchi = c(seq(0.1, 2.9, by = 0.1), 3:10),
  wtr_thresh = c(11, 25), irr_thresh = c(0.0762, 0.6476))
}
\arguments{
\item{lakeid}{ID of lake to run}

\item{secchi}{Vector of secchi values to run model over}

\item{wtr_thresh}{The temperature thresholds in degC (defaults from Lester et al 2004)}

\item{irr_thresh}{The light intensity thresholds in W/m^2 (defaults from Lester et al 2004)}
}
\description{
This runs GLM using a series of different secchi values, and for each calculates the TOHA.
}

