% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_toha_stats.R
\name{calc_toha_stats}
\alias{calc_toha_stats}
\title{Calc standard stats on TOHA for a given lake}
\usage{
calc_toha_stats(lakeid, width_thresh = 0.02, secchi = c(seq(0.1, 2.9, by =
  0.1), 3:10), wtr_thresh = c(11, 25), irr_thresh = c(0.0762, 0.6476))
}
\arguments{
\item{lakeid}{ID of lake to run}

\item{width_thresh}{The fraction of benthic area threshold across which to find the width of the peak (1 is 100%)}

\item{secchi}{Vector of secchi values to run model over}

\item{wtr_thresh}{The temperature thresholds in degC (defaults from Lester et al 2004)}

\item{irr_thresh}{The light intensity thresholds in W/m^2 (defaults from Lester et al 2004)}
}
