% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chained.habitat.calc.kevin.R
\name{chained.habitat.calc.kevin}
\alias{chained.habitat.calc.kevin}
\title{Chained habitat out function for Krose metrics}
\usage{
chained.habitat.calc.kevin(run.path, output.path = NULL, lakeid)
}
\arguments{
\item{lakeid}{}
}
\description{
Calculates all the different metrics requested by Krose
}

