<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
% Please edit documentation in R/comb_output_table.R
\name{comb_output_table}
\alias{comb_output_table}
\title{Combine file-based table output}
\usage{
comb_output_table(pattern, ...)
}
\arguments{
\item{pattern}{Pattern to be passed to \code{\link{Sys.glob}}}

\item{...}{Additional parameters passed to \code{\link{read.table}}}
}
\description{
Attempts to load all files matching a certain pattern and combine
them vertically (rbind). Returns the result. If no files match the
pattern, then an empty data.frame is returned.
}
\details{

}
\author{
Luke Winslow
}

