% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_output_data.R
\name{combine_output_data}
\alias{combine_output_data}
\title{combine full sim run output data}
\usage{
combine_output_data(sim, path)
}
\description{
Combines all the individual compute node model files into 
a few files for the whole simulation
}

