% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_irr_day_cycle.R
\name{create_irr_day_cycle}
\alias{create_irr_day_cycle}
\title{Estimate a full irradiance diurnal cycle}
\usage{
create_irr_day_cycle(lat, lon, dates = as.POSIXct("1990-01-01"),
  irr_mean = 1, by = "hour")
}
\arguments{
\item{lat}{Latitude in degrees}

\item{lon}{Longitude in degrees}

\item{dates}{A vector of dates to generate the diurnal cycles, will be extrapolated to hour, min, etc based on input to \code{by}}

\item{irr_mean}{The mean daily irradiance to match in the diurnal cycle (arbitrary units), must be the same length as \code{dates}}

\item{by}{Timestep to generate sequence of times, see \link{seq.POSIXt} for options}
}
\description{
Estimate a full irradiance diurnal cycle
}

