% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{download_helper}
\alias{download_helper}
\title{Download helper to handle retries}
\usage{
download_helper(url, dest, retries = 3)
}
\arguments{
\item{url}{Full file URL}

\item{dest}{Desired file destination}

\item{retries}{Number of download retries in case of failure (default:3)}
}
\value{
Returns TRUE if download successful, 1 if download fails
}
\description{
Attempts to download file to destination, retries if
the download fails for whatever reason
}
