% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_modification_functions.R
\name{driver_add_rain}
\alias{driver_add_rain}
\title{Add rain to specific month}
\usage{
driver_add_rain(drivers, months = 7:9, rain_add = 1)
}
\arguments{
\item{drivers}{A data.frame with driver data}

\item{months}{The numeric months to add rain (Defaults to summer months 7-9)}

\item{rain_add}{Amount of rain to add across the months in meters}
}
\details{
This is used to artificially add rain to a met file during 
specific months (when it will impact heat budget the least).
}

