% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/driver_modification_functions.R
\name{driver_nldas_debias_airt_sw}
\alias{driver_nldas_debias_airt_sw}
\title{Debias a driver dataset using NLDAS}
\usage{
driver_nldas_debias_airt_sw(drivers, nldas)
}
\description{
Debiases airT and SW based on a linear model of driver to NLDAS
airT and a simple offset with SW
}

