% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_modification_functions.R
\name{driver_nldas_wind_debias}
\alias{driver_nldas_wind_debias}
\title{Debias NLDAS wind}
\usage{
driver_nldas_wind_debias(drivers, ...)
}
\arguments{
\item{drivers}{A GLM drivers data.frame}
}
\description{
Debias NLDAS wind based on known post-2001 NLDAS/NARR 
issue
}

