% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fix_ice_on_off.R
\name{fix_ice_on_off}
\alias{fix_ice_on_off}
\title{Fix ice on/off for no-ice}
\usage{
fix_ice_on_off(on_off, wtr, hypso)
}
\arguments{
\item{on_off}{data.frame of ice on-off dates from \link{get_ice_onoff}.}

\item{wtr}{Water temperature data.frame from \link[glmtools]{get_temp}}

\item{hypso}{Hypsography data.frame in rLakeAnalyzer format \link[rLakeAnalyzer]{load.bathy}}
}
\description{
Calculating metrics on lakes that sometimes, but do not consistently freeze
is a bear, especially with future projections which forecast more frequent
no-freeze years. This function fills in missing ice on/off dates with
the date of minimum average lake temp. Seems to be a reasonable enough
choice. There is no great one.
}

