% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_driver_nhd.R
\name{get_driver_index}
\alias{get_driver_index}
\title{Return the driver file index table}
\usage{
get_driver_index(driver_name, loc_cache = TRUE)
}
\arguments{
\item{driver_name}{The driver source name, options ('NLDAS', 'CM2.0', 'GENMOM', 'ECHAM5')}

\item{loc_cache}{Should the locally cached version of the driver file be used if available.}
}
\description{
Accesses and returns the driver file index. Can be used to get list of
all available driver data for a given driver
}
\examples{
unique(get_driver_index('NLDAS')$id)
}

