<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
% Please edit documentation in R/get_driver_path.R
\name{get_driver_path}
\alias{get_driver_path}
\title{Return driver file location for given lake}
\usage{
get_driver_path(fname)
}
\arguments{
\item{fname}{Name of the file, generally based on the unique site_id}
}
\description{
This functoin returns a path to the driver file for a lake. Retrieve and cache
the file from Sciencebase if necessary
}
\author{
Luke Winslow
}

