% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_driver_path.R
\name{get_driver_wbic}
\alias{get_driver_wbic}
\title{Internal function for WBIC drivers}
\usage{
get_driver_wbic(fname, driver_name, loc_cache, timestep)
}
\description{
Old style download for legacy WBIC 
driver files
}
\keyword{internal}

