% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_ice_onoff}
\alias{get_ice_onoff}
\title{Get Ice On/Off Dates}
\usage{
get_ice_onoff(ice, wtr)
}
\arguments{
\item{ice}{data.frame of ice data, from \code{get_ice} function}

\item{wtr}{data.frame of water temp data, from \code{get_temp} function}
}
\value{
data.frame with columns \code{c('year', 'on', 'off')}
}
\description{
Function to estimate the ice on/off dates from a timeseries of water temperature
and ice data.
}
\details{

}

