% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_kd_year.R
\name{get_kd_year}
\alias{get_kd_year}
\title{Get light attenuation coefficient for a lake through time}
\usage{
get_kd_year(site_id, years)
}
\arguments{
\item{site_id}{The character ID for the requested data}

\item{years}{Integer vector of years for which you want Kd values}
}
\value{
Vector of Kd values with the same length as \code{years}
}
\description{
Returns the annual average light attenuation coefficient for a lake with the given ID
}
\details{
TODO
}
\examples{
#Get and plot Kds for a lake over time
years = 1970:2012
kds   = get_kd_year('6100', years)
plot(years, kds, type='o')
}
\author{
Luke Winslow, Jordan Read
}

