% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_met_drivers_gdp}
\alias{get_met_drivers_gdp}
\title{Submit and retrieve GDP processing request for met drivers}
\usage{
get_met_drivers_gdp(dest_path = ".", service = "regclim:ncep",
  start = as.POSIXct("1990-01-01"), end = as.POSIXct("1990-01-03"))
}
\arguments{
\item{dest_path}{Folder path to store downloaded files when finished}

\item{service}{The source data thredds service}

\item{start}{Start date for met data timeseries as \link{POSIXct}}

\item{end}{End date for met data timeseries as \link{POSIXct}}
}
\description{
Available services:\cr
\code{nldas}\cr
\code{regclim:echam5}\cr
\code{regclim:genmom}\cr
\code{regclim:cm2.0}\cr
\code{regclim:ncep}
}
\details{

}
\author{
Luke Winslow, Jordan Read
}

