<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
% Please edit documentation in R/get_wtr_chained.R
\name{get_wtr_chained}
\alias{get_wtr_chained}
\title{Get full WTR data frame for chained run}
\usage{
get_wtr_chained(run.dir)
}
\description{

}

