<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
% Please edit documentation in R/lake_attribute_functions.R
\name{getArea}
\alias{getArea}
\title{Get lake surface area}
\usage{
getArea(site_id)
}
\arguments{
\item{site_id}{The character ID for the requested data (can be vector of ids)}
}
\value{
Lake surface areas in meters^2. NA if no value available
}
\description{
Returns the surface area for a lake with given ID
}
\details{
Looks for given site_ids and returns values if availalbe. All site_ids
are coerced to character regardless of input type.
}
\examples{
#Multiple values
getArea(c('10000', '6100'))

#this should return NA
getArea('asdf')
}
\author{
Jordan Read, Luke Winslow
}

