% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCanopy}
\alias{getCanopy}
\title{Get surrounding canopy height for a given lake}
\usage{
getCanopy(site_id, default.if.null = FALSE, method = "aster")
}
\arguments{
\item{site_id}{The character ID for the requested data}

\item{default.if.null}{Default value to return if canopy height is unknown}

\item{method}{Canopy height estimation method [aster or landcover]}
}
\value{
Canopy height above lake surface level in meters
}
\description{
Returns the surrounding canopy height for a lake with the given ID
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

