% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getClarity}
\alias{getClarity}
\title{Get light attenuation coefficient for a given lake}
\usage{
getClarity(WBIC, default.if.null = FALSE)
}
\arguments{
\item{default.if.null}{boolean indicating if default Kd should be used if lake has no observations}

\item{site_id}{The character ID for the requested data}
}
\value{
Light attenuation coefficient in m^-1
}
\description{
Returns the light attenuation coefficient for a lake with the given ID
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

