% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_attribute_functions.R
\name{getIceOn}
\alias{getIceOn}
\title{Get the ice on date for a given lake}
\usage{
getIceOn(WBIC, year, fname = "empirical.ice.tsv")
}
\arguments{
\item{year}{The year for which you need the ice-on date}

\item{site_id}{The character ID for the requested data}
}
\value{
Max observed depth in meters
}
\description{
Get the ice on date for a given lake id and year
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}
