% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLatLon}
\alias{getLatLon}
\title{Get latitude and longitude for a given lake}
\usage{
getLatLon(WBIC)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Lat/lon on the WGS84 datum
}
\description{
Get the center lat/lon of a lake with given ID
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

