<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
% Please edit documentation in R/lake_attribute_functions.R
\name{getPerim}
\alias{getPerim}
\title{Get perimeter for a given lake}
\usage{
getPerim(WBIC)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Perimeter in meters
}
\description{
Get the perimeter of a lake with given ID
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

