<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
% Please edit documentation in R/lake_attribute_functions.R
\name{getResidenceTime}
\alias{getResidenceTime}
\title{Get estimated lake residence time}
\usage{
getResidenceTime(WBIC, default.if.null = FALSE)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Estimated residence time in days
}
\description{
Returns the estimated residence time for a lake with the given ID
}
\details{
TODO
}
\examples{
TODO
}
\author{
Luke Winslow, Jordan Read
}
\references{
TODO Data source needed
}

