% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_attribute_functions.R
\name{getScenarioKd}
\alias{getScenarioKd}
\title{Get light attenuation based on long-term trend scenario for a given lake}
\usage{
getScenarioKd(WBIC, years, year.1 = 1979, year.2 = 2011, trend = 0,
  default.if.null = FALSE)
}
\arguments{
\item{years}{Numeric year vector across which to calculate Kd values}

\item{year.1}{The numeric year bounds of the averaging
(i.e., values outside of this range will not be used)}

\item{trend}{The percentage of Secchi increase per year. 
Positive is increasing secchi (e.g., 0.94 is 0.94% increase per year)}

\item{default.if.null}{boolean indicating if default Kd should be used if lake has no observations}

\item{site_id}{The character ID for the requested data}
}
\value{
light attenuation coefficient in m^-1
}
\description{
Calculate the long-term values of light attenuation for a given lake based on
changing scenario
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}
