% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lake_attribute_functions.R
\name{getSDF}
\alias{getSDF}
\title{Get shoreline development factor for a given lake}
\usage{
getSDF(WBIC)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Return a SDF (between 1 and infinity)
}
\description{
Get the shoreline development factor (SDF) of a lake with given ID
}
\details{
The ratio of a a lake's observed perimeter divided by the perimeter of a
circle with the same area as the lake. Cannot be less than 1.
}
\author{
Luke Winslow, Jordan Read
}

