% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getZmax}
\alias{getZmax}
\title{Get max depth for a given lake}
\usage{
getZmax(WBIC)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Max observed depth in meters
}
\description{
Get the max depth for a given lake id
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

