% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opti_thermal_habitat.R
\name{interp_hypso}
\alias{interp_hypso}
\title{Linear interpolate hypsometry}
\usage{
interp_hypso(hypso, dz = 0.1, force_zero_area = TRUE)
}
\arguments{
\item{hypso}{Hypsography data.frame with columns \code{depth} (meters, positive downward) and \code{area} (m^2)}

\item{dz}{Requested resample depth interval}

\item{force_zero_area}{Force the hypsography profile to keep a zero-area depth, even if the input did not have
a zero-area depth, or if the zero area depth does not fall perfectly on the \code{dz} interval}
}
\value{
A hypso data.frame with columns \code{depth and area}
}
\description{
Interpolate bathymetry to a specified depth interval. Includes
options to ensure validity of profile (force zero area layer, monotonically
decreasing area)
}
\examples{
hypso = getBathy('439800')
new_hypso = interp_hypso(hypso, dz=0.1)

plot(hypso$area, hypso$depth, ylim=rev(range(new_hypso$depth)), type='o')
lines(new_hypso$area, new_hypso$depth, lty=2, col='red')
}

