% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necsc_thermal_metrics_core.R
\name{necsc_thermal_metrics_core}
\alias{necsc_thermal_metrics_core}
\title{Calculate core NECSC thermal metrics}
\usage{
necsc_thermal_metrics_core(run.path, site_id)
}
\arguments{
\item{run.path}{The directory where the output.nc and glm2.nml file can be found for the 
simulation run.}

\item{site_id}{The full ID of the lake site as a character string (e.g., 'nhd_13344210')}
}
\value{
A data frame with all metrics organized by year
}
\description{
Calculates the list of core metrics decided by the team. 
See inst/supporting_files/necsc_thermal_metrics_core_outputs.xlsx for details
}

