% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_run_chained_glm.R
\name{prep_run_chained_glm}
\alias{prep_run_chained_glm}
\title{Prepare and run chained model based on unique site_id}
\usage{
prep_run_chained_glm(site_id, path, start = as.POSIXct("2008-04-01"),
  end = as.POSIXct("2008-06-01"), nml_args = NULL)
}
\arguments{
\item{path}{}

\item{nml_args}{}
}
\description{
Pulls in the required information to do a simple GLM 
run. Grabs driver data from Sciencbase and attribute data
from the internal package DB.
}
\examples{
dir.create('~/test')
dir.create(to_run[i])

prep_run_chained_glm(to_run[i], path=to_run[i], 
								 start=as.POSIXct('1979-01-01'), 
								 end=as.POSIXct('2011-12-31'))
								 

}

