% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_run_chained_glm_kd.R
\name{prep_run_chained_glm_kd}
\alias{prep_run_chained_glm_kd}
\title{Prepare and run chained model with Kd scenario}
\usage{
prep_run_chained_glm_kd(site_id, path, years, kd = getClarity(site_id,
  default.if.null = TRUE), nml_args = NULL, verbose = FALSE,
  ice_src = "empirical.ice.tsv")
}
\arguments{
\item{site_id}{}

\item{path}{Directy to start and run model}

\item{years}{Vector of numeric years to simulate.}

\item{kd}{Directly supplied kd value or values. Must be length 1 (repeated across all years) or same length as number of years simulated}

\item{nml_args}{Arguments for nml file to override defaults}
}
\description{
Pulls in the required information to do a simple GLM 
run. Grabs driver data from SB and attribute data
from the internal package DB. Uses supplied Kd values
}
\examples{
dir.create('~/test')
dir.create(to_run[i])

prep_run_chained_glm(to_run[i], path=to_run[i], 
								 start=as.POSIXct('1979-01-01'), 
								 end=as.POSIXct('2011-12-31'))
								 

}

