% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prep_run_chained_glm_secchi}
\alias{prep_run_chained_glm_secchi}
\title{Prepare and run chained model with Kd scenario}
\usage{
prep_run_chained_glm_secchi(site_id, path, start = as.POSIXct("2008-04-01"),
  end = as.POSIXct("2008-06-01"), secchi_trend = 0, nml_args = NULL)
}
\arguments{
\item{site_id}{}

\item{path}{}

\item{nml_args}{}

\item{kd_trend}{The percentage of Secchi increase per year. Positive is increasing secchi (e.g., 0.94 is 0.94% increasing secchi depth per year)}
}
\description{
Pulls in the required information to do a simple GLM
run. Grabs driver data from Sciencbase and attribute data
from the internal package DB.
}
\examples{
dir.create('~/test')
	dir.create(to_run[i])

	prep_run_chained_glm(to_run[i], path=to_run[i],
									 start=as.POSIXct('1979-01-01'),
									 end=as.POSIXct('2011-12-31'))
}

