% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_necsc_lake.R
\name{run_necsc_lake}
\alias{run_necsc_lake}
\title{Large wrapper function NECSC mod run}
\usage{
run_necsc_lake(site_id = NA, driver_name, out_dir)
}
\description{
Runs a single NECSC lake given the default configuration for 
both NLDAS and Notaro drivers
}
