% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_archive_model.R
\name{sb_archive_model}
\alias{sb_archive_model}
\title{Upload and archive model to ScienceBase}
\usage{
sb_archive_model(path, sb_root, sbuser, sbpass)
}
\arguments{
\item{path}{Path to folder containing sim files}

\item{sb_root}{SB root item ID to create new items under}

\item{sbuser}{SB user name}

\item{sbpass}{SB password}
}
\description{
Creates relevant child items a
}
