% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sens_seasonal.R
\name{sens_seasonal_site}
\alias{sens_seasonal_site}
\title{Seasonal, cross-site Sen's slope}
\usage{
sens_seasonal_site(times, data, season_i, sites_i)
}
\arguments{
\item{times}{Numeric vector of times (year for example)}

\item{data}{Numeric vector of observation values}

\item{season_i}{Numeric vector of season identifiers (example: week number)}

\item{sites_i}{Optional numeric of site identifiers (can be numeric or string)}
}
\description{
This outputs all paired-value slopes for a seasonal/cross-site Sen's
slope estimator. To get the slope, calculate the median value of all
paired-value slopes.
}

