% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_driver_nhd.R
\name{set_driver_url}
\alias{set_driver_url}
\title{Set driver URL}
\usage{
set_driver_url(url)
}
\arguments{
\item{url}{New base URL to set}
}
\description{
Sets the default URL to access driver data.
}

