% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_var_notaro.R
\name{summarize_notaro}
\alias{summarize_notaro}
\title{Summary stats for Notaro}
\usage{
summarize_notaro(nc.file)
}
\arguments{
\item{nc.file}{path to nc file}
}
\value{
data.frame with summary statistics
}

