% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_error_log.R
\name{write_error_log}
\alias{write_error_log}
\title{Simple error log writing function}
\usage{
write_error_log(error_list, ids, fname)
}
\arguments{
\item{error_list}{List of error text output}

\item{ids}{List of Lake IDs, same length as error_list}

\item{fname}{File  name to write log output}
}
\description{

}

