<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chained.habitat.calc.kevin.R
\name{chained.habitat.calc.kevin}
\alias{chained.habitat.calc.kevin}
\title{Chained habitat out functoin for Krose metrics}
\usage{
chained.habitat.calc.kevin(run.path, output.path = NULL, lakeid)
}
\description{
Calculates all the different metrics requested by Krose
}

=======
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chained.habitat.calc.kevin.R
\name{chained.habitat.calc.kevin}
\alias{chained.habitat.calc.kevin}
\title{Chained habitat out functoin for Krose metrics}
\usage{
chained.habitat.calc.kevin(run.path, output.path = NULL, lakeid)
}
\description{
Calculates all the different metrics requested by Krose
}

>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
