% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chained.habitat.calc.kevin}
\alias{chained.habitat.calc.kevin}
\title{Chained habitat out functoin for Krose metrics}
\usage{
chained.habitat.calc.kevin(run.path, output.path = NULL, lakeid)
}
\description{
Chained habitat out functoin for Krose metrics
}

