% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_driver_path.R
\name{get_driver_path}
\alias{get_driver_path}
\title{Return driver file location for given lake}
\usage{
get_driver_path(fname, driver_zip = "drivers.zip")
}
\arguments{
\item{fname}{Name of the file, generally based on the unique site_id}

\item{driver_zip}{The path to the driver zip file. Assumed to be in working directory and called 'drivers.zip'}
}
\description{
This functoin returns a path to the driver file for a lake. Retrieve and cache
the file from Sciencebase if necessary
}
\author{
Luke Winslow
}

