% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_driver_path.R
\name{get_driver_path}
\alias{get_driver_path}
\title{Return driver file location for given lake}
\usage{
get_driver_path(id, driver_name = "NLDAS", loc_cache = TRUE,
  timestep = "daily", ...)
}
\arguments{
\item{id}{Site id, prepended with id type followed by _ (e.g., 'nhd_' or 'wbic_')}

\item{driver_name}{The driver source name, options ('NLDAS', 'CM2.0', 'GENMOM', 'ECHAM5')}

\item{loc_cache}{Should the locally cached version of the driver file be used if available.}

\item{timestep}{Requested timestep, "hourly" or "daily" (default) accepted}

\item{...}{Extra parameters are ignored}
}
\description{
This functoin returns a path to the driver file for a lake. Retrieve and cache
the file from Sciencebase if necessary
}
\examples{
get_driver_path('nhd_120052892')
get_driver_path('nhd_120052892', 'ECHAM5')
get_driver_path('nhd_120052892', 'CM2.0')


get_driver_path('WBIC_1881900')

}
\author{
Luke Winslow
}

