% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getArea}
\alias{getArea}
\title{Get lake surface area}
\usage{
getArea(site_id)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Lake surface area in meters^2
}
\description{
Returns the surface area for a lake with given ID
}
\details{
TODO
}
\examples{
TODO
}
\author{
Jordan Read, Luke Winslow
}

