<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
% Please edit documentation in R/lake_attribute_functions.R
\name{getBathy}
\alias{getBathy}
\title{Get Hypsometry for a given lake}
\usage{
getBathy(site_id)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Data frame with columns \code{depth} and \code{area}
}
\description{
Returns the hypsometry profile for a lake with the given ID
}
\author{
Luke Winslow, Jordan Read
}

