% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_attribute_functions.R
\name{getCD}
\alias{getCD}
\title{Get coefficient of wind drag for a given lake or given wind sheltering coefficient}
\usage{
getCD(WBIC = NULL, Wstr = NULL)
}
\arguments{
\item{Wstr}{The wind sheltering coefficient}

\item{site_id}{The character ID for the requested data}
}
\value{
Coefficient of wind drag
}
\description{
Get coefficient of wind drag for a lake with a given ID or a supplied wind sheltering coefficient
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

