% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getElevation}
\alias{getElevation}
\title{Get surface elevation for a given lake}
\usage{
getElevation(WBIC)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Elevation in meters
}
\description{
Get the elevation of a lake with given ID
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

