% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_attribute_functions.R
\name{getElevation}
\alias{getElevation}
\title{Get surface elevation for a given lake}
\usage{
getElevation(WBIC)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Elevation in meters
}
\description{
Get the elevation of a lake with given ID
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

