% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lake_attribute_functions.R
\name{getIceOff}
\alias{getIceOff}
\title{Get the ice off date for a given lake}
\usage{
getIceOff(site_id, year, fname = "empirical.ice.tsv")
}
\arguments{
\item{site_id}{The character ID for the requested data}

\item{year}{The year for which you need the ice-off date}
}
\value{
A vector of ice off dates (as character %Y-%m-%d) with the same length as supplied \code{site_id}
}
\description{
Get the ice off date for a given lake id and year
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

