% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIceOn}
\alias{getIceOn}
\title{Get the ice on date for a given lake}
\usage{
getIceOn(WBIC, year)
}
\arguments{
\item{year}{The year for which you need the ice-on date}

\item{site_id}{The character ID for the requested data}
}
\value{
Max observed depth in meters
}
\description{
Get the ice on date for a given lake id and year
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

