% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_attribute_functions.R
\name{getLatLon}
\alias{getLatLon}
\title{Get latitude and longitude for a given lake}
\usage{
getLatLon(site_id)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Lat/lon on the WGS84 datum as a list
}
\description{
Get the center lat/lon of a lake with given ID
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

