% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_attribute_functions.R
\name{getResidenceTime}
\alias{getResidenceTime}
\title{Get estimated lake residence time}
\usage{
getResidenceTime(WBIC, default.if.null = FALSE)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Estimated residence time in days
}
\description{
Returns the estimated residence time for a lake with the given ID
}
\details{
TODO
}
\examples{
TODO

}
\references{
TODO Data source needed
}
\author{
Luke Winslow, Jordan Read
}
