<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 5ae187e09e693d8c96deb7b89dc0bfd5b01538e1
% Please edit documentation in R/lake_attribute_functions.R
\name{getWstr}
\alias{getWstr}
\title{Get wind sheltering coefficient for a given lake}
\usage{
getWstr(WBIC, method = "Markfort", canopy = NULL)
}
\arguments{
\item{method}{The desired calculation method one of c('Markfort', 'Hondzo', )}

\item{site_id}{The character ID for the requested data}
}
\value{
The wind sheltering coefficient (between 0 and 1)
}
\description{
Get the wind sheltering coefficient of a lake with given ID
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

