% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getZmean}
\alias{getZmean}
\title{Get mean depth for a given lake}
\usage{
getZmean(WBIC)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Mean calculated depth in meters
}
\description{
Get the mean depth for a given lake id
}
\details{
TODO
}
\author{
Luke Winslow, Jordan Read
}

