% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/met_check_download_gdp.R
\name{met_check_download_gdp}
\alias{met_check_download_gdp}
\title{Check met data status and download if finished GDP processing}
\usage{
met_check_download_gdp(gdp_list, dest_path = ".")
}
\arguments{
\item{gdp_list}{List of GDP process objects from \code{\link{geoknife}}}

\item{dest_path}{Destiation directory for GDP output to be saved}
}
\value{
The list with finished entries removed
}
\description{

}

