% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opti_thermal_habitat.R
\name{opti_thermal_habitat}
\alias{opti_thermal_habitat}
\title{Calculate optical thermal habitat using temp and light thresholds}
\usage{
opti_thermal_habitat(nc_file, nml_file, irr_thresh = c(0, 2000),
  wtr_thresh = c(0, 25), area_type = "benthic")
}
\value{
data.frame with three columns. opti_hab, therm_hab, opti_therm_hab
for areas of each habitat type (with opti_therm_hab being the overlap of both)
}
\description{
Calculate optical thermal habitat using temp and light thresholds
}

