% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/populate_base_lake_nml.R
\name{populate_base_lake_nml}
\alias{populate_base_lake_nml}
\title{Create basic NML file object for a lake}
\usage{
populate_base_lake_nml(site_id)
}
\arguments{
\item{site_id}{Unique lake ID}
}
\description{
This creates an NML object and pre-fills it with
key parameters for the site_id
}

