% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prep_run_chained_glm}
\alias{prep_run_chained_glm}
\title{Prepare and run chained model based on unique site_id}
\usage{
prep_run_chained_glm(site_id, path)
}
\description{
Pulls in the required information to do a simple GLM
run. Grabs driver data from Sciencbase and attribute data
from the internal package DB.
}
\examples{
dir.create('~/test')
prep_run_chained_glm('10000', '~/test')
}

