% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prep_run_glm.R
\name{prep_run_glm}
\alias{prep_run_glm}
\title{Prepare and run model based on unique site_id}
\usage{
prep_run_glm(site_id, path, start = as.POSIXct("2008-04-01"),
  end = as.POSIXct("2008-06-01"), nml_args = NULL)
}
\description{
Pulls in the required information to do a simple GLM
run. Grabs driver data from Sciencbase and
}
\examples{
dir.create('~/test')
prep_run_chained_glm('10000', '~/test')
}

